// src/events/enums/event.enums.ts
export enum EventStatus {
    DRAFT = 'DRAFT',
    PUBLISHED = 'PUBLISHED',
    CANCELLED = 'CANCELLED',
    COMPLETED = 'COMPLETED',
}

export enum EventType {
    SEMINAR = 'SEMINAR',
    WORKSHOP = 'WORKSHOP',
    HACKATHON = 'HACKATHON',
    WEBINAR = 'WEBINAR',
    OTHER = 'OTHER',
}

export enum EventMode {
    ONLINE = 'ONLINE',
    IN_PERSON = 'IN_PERSON',
    HYBRID = 'HYBRID',
}